refactor(accesslog): Split NCSA and JSON into dedicated files (#44)
* refactor(accesslog): Split NCSA and JSON into dedicated files

Extract the NCSA-specific format helpers, per-token appendFuncs map,
and parseNCSAFormat routine out of accesslog.go into a new ncsa.go.
Add a matching newNCSAAccessLog constructor that mirrors the existing
newJSONAccessLog shape.

accesslog.go now holds only the pipeline skeleton (bufio.Writer,
sync.Pool, flushLoop) and a small newAccessLog dispatcher that
forwards to either newNCSAAccessLog or newJSONAccessLog based on
cfg.AccessLog.Format. The shared pipeline initialization is factored
into newSkeleton / startFlushLoop so both constructors can reuse it
without duplicating state setup.

No behavior change; 0 allocs/op maintained for Common, Combined, and
JSON. File layout now matches the CLAUDE.md foo.go / foo_test.go rule:

  accesslog.go / accesslog_test.go  (pipeline + NewAccessLog)
  ncsa.go      / ncsa_test.go       (NCSA format)
  json.go      / json_test.go       (JSON format)

* test(accesslog): Speed up tests by bypassing the flush ticker

The NCSA and JSON integration tests were each spending ~1s per case
waiting for the default 1-second flush interval, because the 4096-byte
bufio.Writer never filled during the tiny test writes. Set
BufferSize: 1 in the per-case configs so bufio becomes a pass-through
and Log() flushes synchronously.

Also give the NCSA wroteCh a buffer of 1 to match the JSON test. The
old unbuffered channel deadlocked once Log() started writing through
bufio synchronously: Log() holds l.mu during Write, the Delegator
sends on wroteCh before returning, and the test only receives from
wroteCh after Log() returns — causing the background flushLoop to
pile up on l.mu as well.

Suite timing:

  pkg/logger/accesslog:  11.3s -> 3.5s
  pkg/... total:         36.0s -> 23.7s
