+++ pkg/logger/accesslog/accesslog.go
@@ -3,21 +3,12 @@ package accesslog
 import (
 	"bufio"
 	"io"
-	"net"
-	"os"
-	"regexp"
 	"sync"
 	"time"
 
 	"github.com/fasthttpd/fasthttpd/pkg/config"
 	"github.com/fasthttpd/fasthttpd/pkg/logger"
-	"github.com/fasthttpd/fasthttpd/pkg/util"
 	"github.com/valyala/fasthttp"
-)
-
-const (
-	FormatCommon   = `%h %l %u %t "%r" %>s %b`
-	FormatCombined = `%h %l %u %t "%r" %>s %b "%{Referer}i" "%{User-agent}i"`
 )
 
 var (
@@ -52,6 +43,14 @@ func NewAccessLog(cfg config.Config) (AccessLog, error) {
 	return l, nil
 }
 
+// newAccessLog dispatches to the appropriate per-format constructor.
+func newAccessLog(out logger.Rotator, cfg config.Config) (*accessLog, error) {
+	if cfg.AccessLog.Format == FormatJSON {
+		return newJSONAccessLog(out, cfg)
+	}
+	return newNCSAAccessLog(out, cfg)
+}
+
 type accessLog struct {
 	out               logger.Rotator
 	appendLine        appendFunc
@@ -63,13 +62,15 @@ type accessLog struct {
 	closed            bool
 }
 
-func newAccessLog(out logger.Rotator, cfg config.Config) (*accessLog, error) {
+// newSkeleton allocates an *accessLog with the shared pipeline state
+// (bufio.Writer, sync.Pool, done channel). The caller is responsible for
+// setting appendLine before invoking startFlushLoop.
+func newSkeleton(out logger.Rotator, cfg config.Config) *accessLog {
 	bufSize := cfg.AccessLog.BufferSize
 	if bufSize <= 0 {
 		bufSize = 4096
 	}
-
-	al := &accessLog{
+	return &accessLog{
 		out:  out,
 		bw:   bufio.NewWriterSize(out, bufSize),
 		done: make(chan struct{}),
@@ -80,10 +81,18 @@ func newAccessLog(out logger.Rotator, cfg config.Config) (*accessLog, error) {
 			},
 		},
 	}
-	return al.init(cfg)
 }
 
-var formatPattern = regexp.MustCompile(`(%(>|{(.+?)})?([a-zA-Z%])|([^%]+))`)
+// startFlushLoop launches the periodic flush goroutine. Call this once
+// after appendLine has been set.
+func (l *accessLog) startFlushLoop(cfg config.Config) {
+	flushInterval := time.Duration(cfg.AccessLog.FlushInterval) * time.Millisecond
+	if flushInterval <= 0 {
+		flushInterval = time.Second
+	}
+	go l.flushLoop(flushInterval)
+}
+
 var timeNow = func() time.Time { return time.Now() }
 
 // Rotate rotate log stream.
@@ -125,90 +134,6 @@ func (l *accessLog) Close() error {
 		}
 	}
 	return nil
-}
-
-func (l *accessLog) init(cfg config.Config) (*accessLog, error) {
-	format := cfg.AccessLog.Format
-	if format == "" {
-		format = FormatCommon
-	}
-
-	if format == FormatJSON {
-		l.appendLine = appendJSONLog
-	} else {
-		fns, err := l.parseFormat(format, cfg)
-		if err != nil {
-			return nil, err
-		}
-		l.appendLine = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-			for _, fn := range fns {
-				dst = fn(dst, ctx)
-			}
-			return dst
-		}
-	}
-
-	flushInterval := time.Duration(cfg.AccessLog.FlushInterval) * time.Millisecond
-	if flushInterval <= 0 {
-		flushInterval = time.Second
-	}
-
-	go l.flushLoop(flushInterval)
-
-	return l, nil
-}
-
-func (l *accessLog) parseFormat(format string, cfg config.Config) ([]appendFunc, error) {
-	var fns []appendFunc
-	for _, ms := range formatPattern.FindAllStringSubmatch(format, -1) {
-		k := ms[4]
-		switch k {
-		case "":
-			fns = append(fns, newAppendBytes([]byte(ms[0])))
-		case "C":
-			fns = append(fns, newAppendCookie(ms[3]))
-		case "e":
-			fns = append(fns, newAppendEnv(ms[3]))
-		case "i":
-			fns = append(fns, newAppendRequestHeader(ms[3]))
-		case "o":
-			fns = append(fns, newAppendResponseHeader(ms[3]))
-		case "p":
-			if ms[3] == "remote" {
-				fns = append(fns, appendLpRemote)
-			} else {
-				fns = append(fns, appendLp)
-			}
-		case "t":
-			if ms[3] == "" {
-				fns = append(fns, appendLt)
-			} else {
-				fns = append(fns, newAppendStrftime(ms[3]))
-			}
-		case "v":
-			if cfg.Host != "" {
-				fns = append(fns, newAppendBytes([]byte(cfg.Host)))
-			} else {
-				fns = append(fns, appendNil)
-			}
-		case "V":
-			canonicalHost, err := os.Hostname()
-			if err != nil {
-				return nil, err
-			}
-			fns = append(fns, newAppendBytes([]byte(canonicalHost)))
-		case "r":
-			l.collectRequestURI = true
-			fallthrough
-		default:
-			if fn, ok := appendFuncs[k]; ok {
-				fns = append(fns, fn)
-			} else {
-				fns = append(fns, newAppendBytes([]byte("%"+k)))
-			}
-		}
-	}
-	return fns, nil
 }
 
 // Collect stores Request-URI to ctx as UserValue if '%r' is specified in format.
@@ -253,325 +178,9 @@ func (l *accessLog) Log(ctx *fasthttp.RequestCtx) {
 	l.bufPool.Put(bp)
 }
 
-// appendNetAddr appends the string representation of addr to dst without
-// allocating for the common *net.TCPAddr case.
-func appendNetAddr(dst []byte, addr net.Addr) []byte {
-	if ta, ok := addr.(*net.TCPAddr); ok {
-		if ip4 := ta.IP.To4(); ip4 != nil {
-			dst = fasthttp.AppendIPv4(dst, ip4)
-		} else {
-			dst = append(dst, '[')
-			dst = append(dst, ta.IP.String()...)
-			dst = append(dst, ']')
-		}
-		dst = append(dst, ':')
-		dst = fasthttp.AppendUint(dst, ta.Port)
-		return dst
-	}
-	return append(dst, addr.String()...)
-}
-
-// portFromAddr extracts the port from addr without allocating for *net.TCPAddr.
-func portFromAddr(addr net.Addr) int {
-	if ta, ok := addr.(*net.TCPAddr); ok {
-		return ta.Port
-	}
-	return 0
-}
-
-func appendLp(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-	if p := portFromAddr(ctx.LocalAddr()); p > 0 {
-		return fasthttp.AppendUint(dst, p)
-	}
-	return appendNil(dst, nil)
-}
-
-func appendLpRemote(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-	if p := portFromAddr(ctx.RemoteAddr()); p > 0 {
-		return fasthttp.AppendUint(dst, p)
-	}
-	return appendNil(dst, nil)
-}
-
-var ncsaMonths = []byte("JanFebMarAprMayJunJulAugSepOctNovDec")
-
-// appendNCSADate appends NCSA common format of date (dd/MMM/yyyy:hh:mm:ss +-hhmm)
-// to dst and returns the extended dst.
-func appendNCSADate(dst []byte, date time.Time) []byte {
-	m := (date.Month() - 1) * 3
-	_, zo := date.Zone()
-
-	b := []byte{
-		'[',
-		'0', 0,
-		'/',
-		0, 0, 0,
-		'/',
-		'0', '0', '0', 0,
-		':',
-		'0', 0,
-		':',
-		'0', 0,
-		':',
-		'0', 0,
-		' ',
-		0, '0', 0,
-		'0', '0', ']',
-	}
-	util.CopyRightUint(b[1:3], date.Day())
-	copy(b[4:7], ncsaMonths[m:m+3])
-	util.CopyRightUint(b[8:12], date.Year())
-	util.CopyRightUint(b[13:15], date.Hour())
-	util.CopyRightUint(b[16:18], date.Minute())
-	util.CopyRightUint(b[19:21], date.Second())
-	if zo < 0 {
-		b[22] = '-'
-		zo = -zo
-	} else {
-		b[22] = '+'
-	}
-	util.CopyRightUint(b[23:25], zo/(60*60))
-
-	return append(dst, b...)
-}
-
-// appendNCSARequest appends NCSA common format of request (method uri protocol)
-// to dst and returns the extended dst.
-func appendNCSARequest(dst, method, uri, protocol []byte) []byte {
-	dst = append(dst, method...)
-	dst = append(dst, ' ')
-	for _, s := range uri {
-		switch s {
-		case '"', '\\':
-			dst = append(dst, '\\')
-		}
-		dst = append(dst, s)
-	}
-	dst = append(dst, ' ')
-	dst = append(dst, protocol...)
-	return dst
-}
-
+// appendFunc is the signature shared by every per-token NCSA helper and the
+// JSON whole-line writer.
 type appendFunc func(dst []byte, ctx *fasthttp.RequestCtx) []byte
-
-func newAppendBytes(b []byte) appendFunc {
-	return func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return append(dst, b...)
-	}
-}
-
-func newAppendCookie(key string) appendFunc {
-	bytesKey := []byte(key)
-	return func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		if v := ctx.Request.Header.CookieBytes(bytesKey); len(v) > 0 {
-			return append(dst, v...)
-		}
-		return appendNil(dst, nil)
-	}
-}
-
-func newAppendEnv(key string) appendFunc {
-	return func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		if v, ok := os.LookupEnv(key); ok {
-			return append(dst, v...)
-		}
-		return appendNil(dst, nil)
-	}
-}
-
-func newAppendRequestHeader(key string) appendFunc {
-	bytesKey := []byte(key)
-	return func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		if v := ctx.Request.Header.PeekBytes(bytesKey); len(v) > 0 {
-			return append(dst, v...)
-		}
-		return appendNil(dst, nil)
-	}
-}
-
-func newAppendResponseHeader(key string) appendFunc {
-	bytesKey := []byte(key)
-	return func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		if v := ctx.Response.Header.PeekBytes(bytesKey); len(v) > 0 {
-			return append(dst, v...)
-		}
-		return appendNil(dst, nil)
-	}
-}
-
-func newAppendStrftime(format string) appendFunc {
-	s := util.NewStrftime(format)
-	return func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return s.AppendBytes(dst, ctx.Time())
-	}
-}
-
-var (
-	// appendNil appends "-".
-	appendNil = newAppendBytes([]byte{'-'})
-	// appendNil appends "+".
-	appendPlus = newAppendBytes([]byte{'+'})
-	// appendLa appends client IP address of the request.
-	appendLa = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return appendNetAddr(dst, ctx.RemoteAddr())
-	}
-	// appendA appends underlying peer IP address of the connection.
-	appendA = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return appendNetAddr(dst, ctx.LocalAddr())
-	}
-	// appendB appends size of response in bytes, excluding HTTP headers.
-	appendB = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return fasthttp.AppendUint(dst, ctx.Response.Header.ContentLength())
-	}
-	// appendLb appends size of response in bytes, excluding HTTP headers.
-	// In CLF format, i.e. a '-' rather than a 0 when no bytes are sent.
-	appendLb = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		if b := ctx.Response.Header.ContentLength(); b > 0 {
-			return fasthttp.AppendUint(dst, b)
-		}
-		return appendNil(dst, nil)
-	}
-	// appendD appends the time taken to serve the request, in microseconds.
-	appendD = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		t := timeNow().Sub(ctx.Time())
-		return fasthttp.AppendUint(dst, int(t.Microseconds()))
-	}
-	// appendLf appends filename.
-	appendLf = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		// TODO: not absolute path
-		return append(dst, ctx.Path()...)
-	}
-	// appendLh appends remote ip.
-	appendLh = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return fasthttp.AppendIPv4(dst, ctx.RemoteIP())
-	}
-	// appendLk appends number of keepalive requests handled on this connection.
-	appendLk = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return fasthttp.AppendUint(dst, int(ctx.ConnRequestNum()))
-	}
-	// appendH appends the request protocol.
-	appendH = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return append(dst, ctx.Request.Header.Protocol()...)
-	}
-	// appendLm appends the request method.
-	appendLm = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return append(dst, ctx.Method()...)
-	}
-	// appendP appends the process ID of the child that serviced the request.
-	appendP = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return fasthttp.AppendUint(dst, os.Getegid())
-	}
-	// appendLq appends the query string (prepended with a ? if a query string
-	// exists, otherwise an empty string).
-	appendLq = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		if qstr := ctx.URI().QueryString(); len(qstr) > 0 {
-			return append(append(dst, '?'), qstr...)
-		}
-		return appendNil(dst, nil)
-	}
-	// appendLr appends first line of request.
-	appendLr = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		uri := ctx.Request.Header.PeekBytes(HeaderKeyOriginalRequestURI)
-		if len(uri) == 0 {
-			return appendNil(dst, nil)
-		}
-		return appendNCSARequest(dst, ctx.Method(), uri, ctx.Request.Header.Protocol())
-	}
-	// appendL appends the request log ID from the error log (or '-' if
-	// nothing has been logged to the error log for this request).
-	appendL = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return fasthttp.AppendUint(dst, int(ctx.ID()))
-	}
-	// appendLs appends final status. %s and %>s are mapped this function.
-	appendLs = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return fasthttp.AppendUint(dst, ctx.Response.StatusCode())
-	}
-	// appendLt appends time the request was received, in the format
-	// [18/Sep/2011:19:18:28 -0400].
-	appendLt = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return appendNCSADate(dst, ctx.Time())
-	}
-	// appendT appends the time taken to serve the request, in seconds.
-	appendT = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		t := timeNow().Sub(ctx.Time())
-		return fasthttp.AppendUint(dst, int(t.Seconds()))
-	}
-	// appendLu appends remote user if the request was authenticated.
-	appendLu = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		if u := ctx.Request.URI().Username(); len(u) > 0 {
-			return append(dst, u...)
-		}
-		return appendNil(dst, nil)
-	}
-	// appendU appends the URL path requested, not including any query string.
-	appendU = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		return append(dst, ctx.Path()...)
-	}
-	// appendX appends connection status when response is completed:
-	//   X = Connection aborted before the response completed.
-	//   + = Connection may be kept alive after the response is sent.
-	//   - = Connection will be closed after the response is sent.
-	appendX = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		// TODO: unsupported X
-		if !ctx.Response.ConnectionClose() {
-			return appendPlus(dst, nil)
-		}
-		return appendNil(dst, nil)
-	}
-	// appendI appends bytes received, including request and headers.
-	// Cannot be zero.
-	appendI = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		h := len(ctx.Request.Header.RawHeaders())
-		b := len(ctx.Request.Body())
-		return fasthttp.AppendUint(dst, h+b)
-	}
-	// appendO appends bytes sent, including headers. May be zero in rare cases
-	// such as when a request is aborted before a response is sent.
-	appendO = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		h := len(ctx.Response.Header.Header())
-		b := len(ctx.Response.Body())
-		return fasthttp.AppendUint(dst, h+b)
-	}
-	// appendS appends bytes transferred (received and sent), including request
-	//  and headers, cannot be zero. This is the combination of %I and %O.
-	appendS = func(dst []byte, ctx *fasthttp.RequestCtx) []byte {
-		h := len(ctx.Request.Header.RawHeaders())
-		b := len(ctx.Request.Body())
-		h += len(ctx.Response.Header.Header())
-		b += len(ctx.Response.Body())
-		return fasthttp.AppendUint(dst, h+b)
-	}
-)
-
-// Refer to https://httpd.apache.org/docs/2.4/en/mod/mod_log_config.html
-var appendFuncs = map[string]appendFunc{
-	"%": newAppendBytes([]byte{'%'}),
-	"a": appendLa,
-	"A": appendA,
-	"B": appendB,
-	"b": appendLb,
-	"D": appendD,
-	"f": appendLf,
-	"h": appendLh,
-	"H": appendH,
-	"k": appendLk,
-	"l": appendNil, // Unsupported
-	"L": appendL,
-	"m": appendLm,
-	"P": appendP,
-	"q": appendLq,
-	"r": appendLr,
-	"R": appendNil, // Unsupported
-	"s": appendLs,
-	"t": appendLt,
-	"T": appendT,
-	"u": appendLu,
-	"U": appendU,
-	"X": appendX,
-	"I": appendI,
-	"O": appendO,
-	"S": appendS,
-}
 
 type nilAccessLog struct{}
